Moved messageHtmlToComponent into its own file
